backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
pandas==2.1.3
openpyxl==3.1.2
xlsxwriter==3.1.9
python-dotenv==1.0.0
pydantic-settings==2.0.3
bcrypt==4.1.2
aiosqlite==0.19.0